Correciones y metodo de instancias
cambie el valor que retorna iva, de 5 a 0,05 y así y cree le metodo de instancias

+++ ProyFactura/src/main/java/modelo/Exento.java
@@ -11,7 +11,7 @@ package modelo;
  */
 public class Exento extends Producto{
     @Override
-    public int iva() {
+    public double iva() {
         return 0;
     }
